requirements.txt
python-dotenv
openai
flask
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2